e2e: CreateSync: wait for all containers to run
Users of CreateSync assume that all pod containers are running if
pod is in 'Ready' state. This is not the case as described in
https://kubernetes.io/docs/concepts/workloads/pods/pod-lifecycle/#pod-phase
Containers can be also [re]starting.

Adding explicit check for all pod containers to be in 'Running' state
should make CreateSync function to satisfy users assumptions.