improve CRD handling in VolumePopulator test
Test: provisioning should provision storage with any volume data source

During CSI certification test we observed that the test can fail
with a message:
"customresourcedefinitions.apiextensions.k8s.io \"volumepopulators.populator.storage.k8s.io\" already exists"

This is because the test does not consider that this CRD can be already
installed in the cluster.

The test was updated to handle the CRD better by creating it for the
duration of the test and removing it afterward. Otherwise, if the CRD
is already installed, the test will neither create nor remove it.